+++ src/lib/sink/file_sink.hpp
@@ -11,6 +11,11 @@
 
 #include "mutex_sink.hpp"
 
+#include <headcode/logger/sink_factory.hpp>
+
+#include <headcode/url/url.hpp>
+
+#include <map>
 #include <string>
 
 
@@ -60,6 +65,63 @@ namespace headcode::logger {
  */
 class FileSink : public MutexSink {
 
+    /**
+     * @brief   Sink producer instance.
+     */
+    struct Producer : public SinkFactory::Producer {
+
+        /**
+         * @brief   Currently known sinks.
+         */
+        static std::map<std::string, std::shared_ptr<FileSink>> sinks;
+
+        /**
+         * @brief   Synchronizes access to sinks member.
+         */
+        static std::mutex mutex;
+
+        /**
+         * @brief   Creates a sink.
+         * This MAY return already created objects.
+         * @param   url         The URL of the sink to create.
+         * @return  A sink instance.
+         */
+        [[nodiscard]] std::shared_ptr<Sink> Create(std::string const & url) override {
+
+            auto parsed_url = headcode::url::URL{url}.Normalize();
+
+            auto lock = std::unique_lock<std::mutex>(mutex);
+            auto iter = sinks.find(parsed_url.GetURL());
+
+            if (iter == sinks.end()) {
+                auto sink = std::make_shared<FileSink>(parsed_url.GetURL());
+                sinks.emplace(parsed_url.GetURL(), sink);
+                return sink;
+            }
+
+            return iter->second;
+        }
+
+        /**
+         * @brief   Returns a human readable id for the sink producer.
+         * This id is also used to identify the producer within the factory.
+         * @return  A description for the sink producer.
+         */
+        [[nodiscard]] std::string GetId() const override {
+            return "FileSink Producer";
+        }
+
+        /**
+         * @brief   Checks if this producer is capable to create the object.
+         * @brief   url         The URL to match against.
+         * @return  True, if this producer can create Sinks matching the given URL.
+         */
+        [[nodiscard]] bool Match(std::string const & url) const override {
+            auto parsed_url = headcode::url::URL{url}.Normalize();
+            return parsed_url.GetScheme() == "file";
+        }
+    };
+
     std::string filename_;        //!< @brief The name of the file to write to.
 
 public:
@@ -68,6 +130,11 @@ public:
      * @param   file_url        URL of the file to write to.
      */
     explicit FileSink(std::string file_url);
+
+    /**
+     * @brief   Registers a Producer at the Sink Factory.
+     */
+    static void RegisterProducer();
 
 private:
     /**