+++ test/unit/test_version.cpp
@@ -18,8 +18,8 @@ TEST(Version, regular) {
 }
 
 
-TEST(Version, V1_1_0) {
-    std::uint32_t version = MAKE_VERSION(1, 1, 0);
-    EXPECT_EQ(static_cast<unsigned int>((1) << 24 | (1) << 16 | (0)), version);
-    EXPECT_EQ(GetHCSLoggerVersion_1_1_0(), version);
+TEST(Version, V2_0_0) {
+    std::uint32_t version = MAKE_VERSION(2, 0, 0);
+    EXPECT_EQ(static_cast<unsigned int>((2) << 24 | (0) << 16 | (0)), version);
+    EXPECT_EQ(GetHCSLoggerVersion_2_0_0(), version);
 }